Finally add some audio!
Also fix a bug in the actions map where the action would get duplicated
if more than one binding was changed without exiting to the main menu.
I'm not even sure why it seemed to work properly in the first place, I
would have expected the duplication to occur anyway. ¯\_(ツ)_/¯